Add the ability to send messages direct to implementations of LeveledLogger
This change allows us to find implementations of LeveledLogger as part
of the DefaultLogger and send messages directly to the LeveledLogger.
This can power optimisations like the backing implementation deciding to
drop messages based on severity or context values without going through
the performance penalty of augmenting the log messages prior.